# Capitalize regex