--- a/Cargo.toml
+++ b/Cargo.toml
@@ -9,3 +9,4 @@ edition = "2018"
 [dependencies]
 rocket = "0.4.7"
 rocket_contrib = "0.4.7"
+regex = "1.4.5"