--- a/src/main.rs
+++ b/src/main.rs
@@ -13,9 +13,15 @@ fn hello(name: String) -> String {
     format!("Hello, {}", string_utils::capitalize(name.as_str()))
 }
 
+#[get("/hello?wave&<name>")]
+fn wave(name: Option<String>) -> String {
+    name.map(|name| format!("Hi, {}!", string_utils::capitalize(name.as_str())))
+        .unwrap_or_else(|| "Hello!".into())
+}
+
 fn main() {
     rocket::ignite()
-        .mount("/", routes![hello])
+        .mount("/", routes![hello, wave])
         .mount("/res", StaticFiles::from("resources"))
         .launch();
 }