Latest benchmark:

before lazy_static!:
```
test string_utils::tests::bench_capitalize       ... bench:       1,789 ns/iter (+/- 709)
test string_utils::tests::bench_capitalize_regex ... bench:     578,563 ns/iter (+/- 50,093)
```

after lazy_static!:
```
test string_utils::tests::bench_capitalize            ... bench:       2,165 ns/iter (+/- 944)
test string_utils::tests::bench_capitalize_poem       ... bench:      69,385 ns/iter (+/- 38,950)
test string_utils::tests::bench_capitalize_regex      ... bench:       1,949 ns/iter (+/- 707)
test string_utils::tests::bench_capitalize_regex_poem ... bench:      73,476 ns/iter (+/- 13,470)
```

after reworking the non-regex version:

```
test string_utils::tests::bench_capitalize            ... bench:         464 ns/iter (+/- 82)
test string_utils::tests::bench_capitalize_poem       ... bench:      16,750 ns/iter (+/- 1,898)
test string_utils::tests::bench_capitalize_regex      ... bench:       1,894 ns/iter (+/- 232)
test string_utils::tests::bench_capitalize_regex_poem ... bench:      73,428 ns/iter (+/- 6,568)
```